Create an entry point to backdrop_collector
In order to unify the few collectors that we currently have we need a
single entry point to run them.

This script will pull the module to execute from the query config, in a
field called "entrypoint". It will expect a function called main in the
module which it will pass the credentials, backdrop information and the
query as arguments.

I've had to introduce a log directory as part of this commit as when
setting up the logging it will be expected to exist. This means we can
also drop the creation and deletion of the directory from the
logging_setup tests.

Thanks to the entry_points setting in setup.py this new script will be
exposed as backdrop_collector when installed into the system.