Migrate GA realtime collector
In comparison with the full ga collector this has been a fairly easy
migration. I've had to alter the tests a bit as Bucket has become
DataSet and the way to create one has changed. Otherwise straight
forward and tested with govuk realtime.

Unsure as to whether we really need to depend on the google api client
again as gapy does but I've left it in.