Fix the repo name for testing the Travis tagger
The travis auto tagger pushes a release tag to the repository when the
build (on master) is successful.

The script allows you to test it in local mode. To do so, you need to
fake the repo name (which would normally come from the Travis
environment)

This fixes the repo name to be `alphagov/performanceplatform-collector` rather
than `alphagov/stagecraft` and adds a testing commit from the repo,
rather than from the stagecraft repo.