Handle (other) date being sent back by GA
GA has started sending back a row of data which contains (other) for the
day, week, and month fields, which currently causes an exception. This
commit skips the row if this is the case.

We do also need to find out the implications of the (other) data
appearing and if this has implication on the collectors, but this fixes
the immediate problem with the collector.