idMapping should work on any field in document
Previously idMapping would only work over dimensions returned by the GA
query. We have a requirement to use a field that will only exist on the
document as part of the id. This commit moves the id generation logic to
after the munging of the document then uses fields in the doc rather
than off the item.