Make ga core collector lazy
This replaces lists and list comprehensions with generators in the ga
core collector. Combined with the changes in 628036e this should allow
the GA core collector to be much more memory efficient by allowing it to
fetch more pages from GA when it needs and send more chunks to backdrop
when it needs.

Note that if plugins are used all bets are off. I have not looked into
them much but the ones I have would not work.