Remove failing test
This test is failing because it runs in a separate python process, which
is not affected by the patched mock.

This test is testing that a cronjob passed in through stdin is appearing
in the output, which due to previous changes is now dependant on the
machines hostname.

This is is tested in other tests which call generate_crontab directly,
so it is not necessary to duplicate it here