Make filtersets compatible with Google Universal Analytics.
Google Analytics Universal does not support custom variables so we
need to change the variable we are filtering on to search for a
department name.

Instead of hardcoding another variable into the search, allow any
variable to be added to the filterset to get the department name.

The variable change only affects the google analytics contrib
content table collector and is only used by gov uk content.