Rename log files
Now we are on gov.uk infrastructure, there is a convention that
applications log to files production.json.log, and files that have the
extension .log. Although there is nothing in puppet to rotate these log
files, currently, we should rename them to be consistent with other apps
before adding to puppet.