Get stagecraft url from file.
The stagecraft url and api token is required to get the collector config
from the database.

For ease of deployment, the stagecraft url and api token have been added
to the same file as the backdrop url.

The key for the backdrop url has been renamed to distinguish it from the
stagecraft url.