Fix how mutually exclusive arguments are handled.
Argparse adds all command line arguments that have been set to mutually exclusive
to the return object. It then giving all of the keys that didn't appear on the
command line a value of None.

This was causing pp-collector to error when the query flag was passed and a
connection to stagecraft was not available. A connection to stagecraft should
only be attempted if the collector-slug flag is used.