Allow ga trending collector to handle missing data.
In cases where there are not 2 weeks worth of data, errors were thrown
when trying to parse the metrics that were required to calculate the %
change.

This now handles that case, but will unfortunately set the missing weeks
data to the floor.  This will ensure that there is some data available
although it will not be 100% accurate.